+++ packages/author-profile/__tests__/ios/author-profile.test.js
@@ -1,5 +1,0 @@
-import sharedNative from "../shared-native";
-import shared from "../author-profile";
-
-shared();
-sharedNative();