+++ packages/author-profile/__tests__/shared.web.js
@@ -1,0 +1,39 @@
+import {
+  addSerializers,
+  compose,
+  minimaliseTransform,
+  minimalWebTransform,
+  print,
+  replacePropTransform
+} from "@times-components/jest-serializer";
+import { hash } from "@times-components/test-utils";
+import shared from "./shared.base";
+import author from "./fixtures";
+
+export default () => {
+  addSerializers(
+    expect,
+    compose(
+      print,
+      minimalWebTransform,
+      minimaliseTransform(
+        (value, key) =>
+          key === "style" || key === "className" || key === "data-testid"
+      ),
+      replacePropTransform(
+        (value, key) => (key === "emptyStateMessage" ? hash(value) : value)
+      )
+    )
+  );
+
+  const props = {
+    analyticsStream() {},
+    author,
+    onNext() {},
+    onPrev() {},
+    refetch() {},
+    slug: "some-slug"
+  };
+
+  shared(props);
+};