+++ packages/author-profile/src/author-profile-head-prop-types.web.js
@@ -1,0 +1,12 @@
+import {
+  propTypes as basePropTypes,
+  defaultPropTypes as baseDefaultPropTypes
+} from "./author-profile-head-prop-types.base";
+
+export const propTypes = {
+  ...basePropTypes
+};
+
+export const defaultProps = {
+  ...baseDefaultPropTypes
+};