+++ packages/jest-configurator/setup-jest.js
@@ -1,4 +1,9 @@
+/* eslint-disable */
+
 import Adapter from "enzyme-adapter-react-16";
 import Enzyme from "enzyme";
 
 Enzyme.configure({ adapter: new Adapter() });
+
+if (typeof window !== "undefined")
+  window.HTMLCanvasElement.prototype.getContext = () => {};