+++ android-app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.4.19"></a>
+## [0.4.19](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.4.18...@times-components/android-app@0.4.19) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/android-app
+
 <a name="0.4.18"></a>
 ## [0.4.18](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.4.17...@times-components/android-app@0.4.18) (2018-08-02)