+++ packages/context/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.1"></a>
+## [0.1.1](https://github.com/newsuk/times-components/compare/@times-components/context@0.1.0...@times-components/context@0.1.1) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/context
+
 <a name="0.1.0"></a>
 # 0.1.0 (2018-08-01)