+++ packages/provider-test-tools/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.15.6"></a>
+## [0.15.6](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@0.15.5...@times-components/provider-test-tools@0.15.6) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/provider-test-tools
+
 <a name="0.15.5"></a>
 ## [0.15.5](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@0.15.4...@times-components/provider-test-tools@0.15.5) (2018-07-31)