+++ packages/slice/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.10"></a>
+## [2.0.10](https://github.com/newsuk/times-components/compare/@times-components/slice@2.0.9...@times-components/slice@2.0.10) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/slice
+
 <a name="2.0.9"></a>
 ## [2.0.9](https://github.com/newsuk/times-components/compare/@times-components/slice@2.0.8...@times-components/slice@2.0.9) (2018-07-31)