+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.7"></a>
+## [2.1.7](https://github.com/newsuk/times-components/compare/@times-components/styleguide@2.1.6...@times-components/styleguide@2.1.7) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/styleguide
+
 <a name="2.1.6"></a>
 ## [2.1.6](https://github.com/newsuk/times-components/compare/@times-components/styleguide@2.1.5...@times-components/styleguide@2.1.6) (2018-07-31)